pyproject.toml
[tool.poetry]
name = "aws-supabse"
version = "0.1.0"
description = "Image Search with Supabase Vector"
authors = ["thorwebdev <thor@supabase.io>"]
readme = "README.md"
packages = [
    {include = "src/image_search"},
    {include = "src/doc_search"},
]

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.7.1"
vecs = "^0.4.3"
boto3 = "^1.34.59"
python-dotenv = "^1.0.1"
langchain = "^0.2.7"
langchain-community = "^0.2.7"
langgraph = "^0.1.8"
httpx = "^0.27.0"
langchainhub = "^0.1.20"
langchain-aws = "^0.1.10"
pypdf = "^4.2.0"
langchain-experimental = "^0.0.62"

[tool.poetry.scripts]
seed_image = "src.image_search.main:seed"
seed_text = "src.doc_search.main:seed"
search_image = "src.image_search.main:search"
search_text = "src.doc_search.main:main"

[tool.poetry.group.dev.dependencies]
setuptools = "^70.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
